+++ app/models/concerns/soft_delete_fields.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module SoftDeleteFields
+    extend ActiveSupport::Concern
+
+    def soft_delete
+      update(deleted_at: Time.current)
+    end
+
+    def activate_deleted_at
+      self.update(deleted_at: nil)
+    end
+end