+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - Config Rspec for tests
 - Enable pgcrypto and add UUID_v7 support
 - Add default version
+- Plan model and tests
 
 ## Version [0.1.0.beta] - 2025-01-20